Backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
executing==2.0.1
idna==3.6
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
